# api-spec/src/main/resources/openapi/gox.yaml
openapi: 3.0.0
info:
  title: GoX API
  version: 1.0.0
  description: API for managing cars in the GoX car rental system.
servers:
  - url: http://localhost:8080/api
paths:
...
  /bookings:
    post:
      summary: Create a new booking
      operationId: createBooking
      tags:
        - Booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingSummary'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
    get:
      summary: Get all bookings
      operationId: getAllBookings
      tags:
        - Booking
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingSummary'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
  /bookings/{bookingId}:
    parameters:
      - name: bookingId
        in: path
        required: true
        schema:
          type: integer
          format: int64

    get:
      summary: Get booking by ID
      operationId: getBookingById
      tags:
        - Booking
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetail'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    delete:
      summary: Delete a booking
      operationId: deleteBooking
      tags:
        - Booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /bookings/{bookingId}/approve:
    post:
      summary: Admin confirms reservation
      operationId: approveBooking
      tags: [ Booking ]
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /bookings/{bookingId}/cancel:
    post:
      summary: Admin/client cancels the reservation
      operationId: cancelBooking
      tags: [ Booking ]
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /bookings/{bookingId}/complete:
    post:
      summary: Complete a booking, set actual return date, calculate penalty
      operationId: completeBooking
      tags:
        - Booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - actualReturnDate
              properties:
                actualReturnDate:
                  type: string
                  format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingSummary'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /bookings/estimate:
    post:
      summary: Estimate cost of a booking without creating
      operationId: estimateBooking
      tags:
        - Booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingEstimate'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
components:
  schemas:
    Car:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the car.
        brand:
          type: string
          description: Brand of the car.
        model:
          type: string
          description: Model of the car.
        year:
          type: integer
          description: Production year of the car.
        seats:
          type: integer
          description: Number of seats in the car.
        pricePerDay:
          type: number
          format: double
          description: Rental price per day.
        state:
          $ref: '#/components/schemas/CarState'
        category:
          $ref: '#/components/schemas/CarCategory'
        transmission:
          $ref: '#/components/schemas/TransmissionType'
        fuelType:
          $ref: '#/components/schemas/FuelType'
        description:
          type: string
          description: Optional description of the car.
        preview:
          $ref: '#/components/schemas/Photo'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        location:
          $ref: '#/components/schemas/Location'
    CarState:
      type: string
      description: The current state of the car.
      enum:
        - AVAILABLE
        - RENTED
        - UNAVAILABLE
        - IN_TRANSIT
    CarCategory:
      type: string
      enum:
        - SEDAN
        - SUV
        - HATCHBACK
        - COUPE
        - WAGON
        - CROSSOVER
        - SPORT
        - CABRIOLET
        - MUV
...
    TransmissionType:
      type: string
      enum:
        - MANUAL
        - AUTOMATIC
    FuelType:
      type: string
      enum:
        - PETROL
        - DIESEL
        - ELECTRIC
        - HYBRID
...
    UserSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        loyaltyLevel:
          $ref: '#/components/schemas/LoyaltyLevel'
        address:
          type: string
        phone:
          type: string
...
    UserRole:
      type: string
      enum:
        - ADMIN
        - CUSTOMER
      description: Role of the user within the system.
...
    LoyaltyLevel:
      type: string
      enum:
        - STANDARD
        - SILVER
        - GOLD
      description: Loyalty level of the user.
...
    Photo:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        isPreview: { type: boolean }
        url:
          type: string
          description: URL to download raw content of this photo
...
    Location:
      type: object
      properties:
        id:
          type: integer
          format: int64
        city:
          type: string
        street:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    BookingSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        actualReturnDate:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/BookingStatus'
        carId:
          type: integer
          format: int64
        carBrand:
          type: string
        carModel:
          type: string
        userId:
          type: integer
          format: int64
        userName:
          type: string
    BookingDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        actualReturnDate:
          type: string
          format: date-time
          nullable: true
        basePrice:
          type: number
          format: double
        loyaltyDiscount:
          type: number
          format: double
        discountedPrice:
          type: number
          format: double
        totalPrice:
          type: number
          format: double
        transferFee:
          type: number
          format: double
        urgent:
          type: boolean
        penalty:
          type: number
          format: double
        status:
          $ref: '#/components/schemas/BookingStatus'
        user:
          $ref: '#/components/schemas/UserSummary'
        car:
          $ref: '#/components/schemas/Car'
        pickupLocation:
          $ref: '#/components/schemas/Location'
        dropoffLocation:
          $ref: '#/components/schemas/Location'
    BookingCreateRequest:
      type: object
      required:
        - startDate
        - endDate
        - carId
        - pickupLocationId
        - dropoffLocationId
      properties:
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        carId:
          type: integer
          format: int64
        pickupLocationId:
          type: integer
          format: int64
        dropoffLocationId:
          type: integer
          format: int64
    BookingStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - CANCELLED
        - COMPLETED
      description: Status of a Booking
    BookingEstimate:
      type: object
      properties:
        days:
          type: integer
        basePrice:
          type: number
          format: double
        loyaltyDiscount:
          type: number
          format: double
        discountedPrice:
          type: number
          format: double
        urgent:
          type: boolean
        transferFee:
          type: number
          format: double
        totalPrice:
          type: number
          format: double
...